requirements.txt
pdfplumber==0.11.4
google-generativeai==0.8.3
python-telegram-bot==21.6
openpyxl==3.1.5
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.28.1
feedparser==6.0.11
